requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.121.2
fonttools==4.60.1
h11==0.16.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
